requirements.txt
cycler==0.10.0
decorator==4.4.0
kiwisolver==1.1.0
matplotlib==3.1.0
networkx==2.3
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
pkg-resources==0.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scipy==1.3.0
six==1.12.0
